NiesBack/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
psycopg2-binary>=2.9
httpx>=0.27
pydantic>=2.8
pydantic-settings>=2.4
python-dotenv>=1.0
alembic>=1.13
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
# email-validator>=2.1.0.post1
pydantic[email]>=2.7.0
bcrypt==4.1.2